Broke code execution into seperate worker service. Added limits to docker-compose file to prevent abuse
