+++ docker-compose.yml
@@ -3,8 +3,18 @@ version: '3.8'
 services:
   server:
     image: smcgarril/leetgo-server:1.0.0
+    restart: always
     build: ./server
     ports:
       - "8080:8080"
+  worker:
+    image: smcgarril/leetgo-worker:1.0.0
+    restart: always
+    build: ./worker
+    ports:
+      - "8081:8081"
+    pids_limit: 60
+    cpus: 1
+    mem_limit: 150M
