+++ worker/go.mod
@@ -1,0 +1,3 @@
+module github.com/smcgarril/leetgo-worker
+
+go 1.22.5
